+++ npm-lib/src/samples/ProtobufSamplesV3Json.ts
@@ -828,7 +828,7 @@ export const jsonDescriptor = {
                                   "id": 11
                                 },
                                 "framesDropped": {
-                                  "type": "double",
+                                  "type": "int32",
                                   "id": 12
                                 },
                                 "framesPerSecond": {
